requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mysql-connector-python
pymysql
websockets